GIX-994: Transaction Form working
