+++ frontend/src/lib/derived/accounts-list.derived.ts
@@ -2,19 +2,19 @@
  * A derived store that returns the accounts as an array of accounts.
  */
 
-import { isNnsProjectStore } from "$lib/derived/selected-project.derived";
-import { snsProjectAccountsStore } from "$lib/derived/sns/sns-project-accounts.derived";
 import { accountsStore } from "$lib/stores/accounts.store";
 import { derived } from "svelte/store";
+import { isNnsProjectStore } from "./selected-project.derived";
+import { snsProjectAccountsStore } from "./sns/sns-project-accounts.derived";
 
 export const accountsListStore = derived(
   [accountsStore, isNnsProjectStore, snsProjectAccountsStore],
-  ([{ main, subAccounts, hardwareWallets }, isNns, projectAccounts]) => {
+  ([{ main, subAccounts, hardwareWallets }, isNns, snsAccounts]) => {
     if (isNns) {
       return main === undefined
         ? []
         : [main, ...(subAccounts ?? []), ...(hardwareWallets ?? [])];
     }
-    return projectAccounts ?? [];
+    return snsAccounts ?? [];
   }
 );
