+++ frontend/src/lib/utils/accounts.utils.ts
@@ -1,8 +1,10 @@
+import { AppPath } from "$lib/constants/routes.constants";
 import type { AccountsStore } from "$lib/stores/accounts.store";
 import type { Account } from "$lib/types/account";
 import { InsufficientAmountError } from "$lib/types/common.errors";
 import { checkAccountId } from "@dfinity/nns";
 import { Principal } from "@dfinity/principal";
+import { getLastPathDetail, isRoutePath } from "./app-path.utils";
 
 /*
  * Returns the principal's main or hardware account
@@ -101,4 +103,30 @@ export const assertEnoughAccountFunds = ({
  */
 export const mainAccount = (accounts: Account[]): Account | undefined => {
   return accounts.find((account) => account.type === "main");
+};
+
+/*
+ * @param path current route path
+ * @return an object containing either a valid account identifier or undefined if not provided for the wallet route or undefined if another route is currently accessed
+ */
+export const routePathAccountIdentifier = (
+  path: string | undefined
+): { accountIdentifier: string | undefined } | undefined => {
+  if (
+    !isRoutePath({
+      paths: [AppPath.LegacyWallet, AppPath.Wallet],
+      routePath: path,
+    })
+  ) {
+    return undefined;
+  }
+
+  const accountIdentifier: string | undefined = getLastPathDetail(path);
+
+  return {
+    accountIdentifier:
+      accountIdentifier !== undefined && accountIdentifier !== ""
+        ? accountIdentifier
+        : undefined,
+  };
 };
