+++ frontend/src/tests/lib/services/accounts.services.spec.ts
@@ -7,7 +7,6 @@ import {
   getAccountIdentityByPrincipal,
   getAccountTransactions,
   renameSubAccount,
-  routePathAccountIdentifier,
   syncAccounts,
   transferICP,
 } from "$lib/services/accounts.services";
@@ -201,27 +200,6 @@ describe("accounts-services", () => {
       });
 
       spyToastError.mockClear();
-    });
-  });
-
-  describe("details", () => {
-    beforeAll(() => {
-      // Avoid to print errors during test
-      jest.spyOn(console, "error").mockImplementation(() => undefined);
-    });
-    afterAll(() => jest.clearAllMocks());
-
-    it("should get account identifier from valid path", () => {
-      expect(
-        routePathAccountIdentifier(`/#/wallet/${mockMainAccount.identifier}`)
-      ).toEqual({
-        accountIdentifier: mockMainAccount.identifier,
-      });
-    });
-
-    it("should not get account identifier from invalid path", () => {
-      expect(routePathAccountIdentifier("/#/wallet/")).toEqual(undefined);
-      expect(routePathAccountIdentifier(undefined)).toBeUndefined();
     });
   });
