+++ frontend/src/tests/lib/utils/neuron.utils.spec.ts
@@ -50,6 +50,7 @@ import {
   minNeuronSplittable,
   neuronCanBeSplit,
   neuronStake,
+  routePathNeuronId,
   sortNeuronsByCreatedTimestamp,
   topicsToFollow,
   userAuthorizedNeuron,
@@ -1741,6 +1742,26 @@ describe("neuron-utils", () => {
           amount: MIN_NEURON_STAKE / 2 / E8S_PER_ICP - 10,
         })
       ).toBe(false);
+    });
+  });
+
+  describe("routePathNeuronId", () => {
+    beforeAll(() => {
+      // Avoid to print errors during test
+      jest.spyOn(console, "error").mockImplementation(() => undefined);
+    });
+    afterAll(() => jest.clearAllMocks());
+    it("should get neuronId from valid path", async () => {
+      expect(routePathNeuronId("/#/neuron/123")).toBe(BigInt(123));
+      expect(routePathNeuronId("/#/neuron/0")).toBe(BigInt(0));
+    });
+
+    it("should not get neuronId from invalid path", async () => {
+      expect(routePathNeuronId("/#/neuron/")).toBeUndefined();
+      expect(routePathNeuronId("/#/neuron/1.5")).toBeUndefined();
+      expect(routePathNeuronId("/#/neuron/123n")).toBeUndefined();
+      expect(routePathNeuronId("/#/neurons/")).toBeUndefined();
+      expect(routePathNeuronId("/#/accounts/")).toBeUndefined();
     });
   });
 });
